# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Plateforme de restaurants en ligne - Burkina Faso
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user (client or restaurant owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with phone/email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Restaurants
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Platform-wide stats (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
  /admin/restaurants:
    get:
      operationId: listAdminRestaurants
      tags: [admin]
      summary: List all restaurants with owner info (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminRestaurant"
        "403":
          description: Forbidden
  /admin/restaurants/{id}/status:
    patch:
      operationId: setRestaurantStatus
      tags: [admin]
      summary: Approve or suspend a restaurant (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantStatusResult"
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "403":
          description: Forbidden
  /admin/users/{id}/role:
    patch:
      operationId: setUserRole
      tags: [admin]
      summary: Change a user's role (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRoleResult"
        "403":
          description: Forbidden
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all platform orders (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminOrder"
        "403":
          description: Forbidden

  # Stats (restaurant dashboard)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [client, restaurant_owner, admin]
        loyaltyPoints:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - role
        - loyaltyPoints
        - createdAt
    UserRegisterInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 8
        email:
          type: string
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [client, restaurant_owner]
      required:
        - name
        - phone
        - password
        - role
    UserLoginInput:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required:
        - phone
        - password
...
    AdminStats:
      type: object
      properties:
        users:
          type: object
          properties:
            total:
              type: integer
            clients:
              type: integer
            restaurantOwners:
              type: integer
            admins:
              type: integer
          required: [total, clients, restaurantOwners, admins]
        restaurants:
          type: object
          properties:
            total:
              type: integer
            pending:
              type: integer
            approved:
              type: integer
            suspended:
              type: integer
          required: [total, pending, approved, suspended]
        orders:
          type: object
          properties:
            total:
              type: integer
            today:
              type: integer
            byStatus:
              type: object
              additionalProperties:
                type: integer
          required: [total, today, byStatus]
        revenue:
          type: object
          properties:
            total:
              type: number
            today:
              type: number
            commission:
              type: number
          required: [total, today, commission]
      required: [users, restaurants, orders, revenue]
    AdminRestaurant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        city:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, suspended]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        ownerId:
          type: integer
        ownerName:
          type: ["string", "null"]
        ownerPhone:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, type, city, phone, status, reviewCount, ownerId, createdAt]
    RestaurantStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, suspended]
      required: [status]
    RestaurantStatusResult:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, approved, suspended]
      required: [id, status]
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [client, restaurant_owner, admin]
        loyaltyPoints:
          type: integer
        createdAt:
          type: string
      required: [id, name, phone, role, loyaltyPoints, createdAt]
    UserRoleUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [client, restaurant_owner, admin]
      required: [role]
    UserRoleResult:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [client, restaurant_owner, admin]
      required: [id, role]
    AdminOrder:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        restaurantId:
          type: integer
        restaurantName:
          type: string
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivering, delivered, cancelled]
        subtotal:
          type: number
        deliveryFee:
          type: number
        discount:
          type: number
        total:
          type: number
        deliveryCity:
          type: string
        paymentMethod:
          type: string
          enum: [orange_money, moov_money, cash]
        itemCount:
          type: integer
        createdAt:
          type: string
      required: [id, customerId, restaurantId, restaurantName, status, subtotal, deliveryFee, discount, total, deliveryCity, paymentMethod, itemCount, createdAt]