# lib/api-zod/src/generated/types/adminRestaurant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Plateforme de restaurants en ligne - Burkina Faso
 * OpenAPI spec version: 0.1.0
 */
import type { AdminRestaurantStatus } from './adminRestaurantStatus';

export interface AdminRestaurant {
  id: number;
  name: string;
  type: string;
  city: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  status: AdminRestaurantStatus;
  /** @nullable */
  rating?: number | null;
  reviewCount: number;
  ownerId: number;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  ownerPhone?: string | null;
  createdAt: string;
}