# lib/api-zod/src/generated/types/restaurantStatusResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Plateforme de restaurants en ligne - Burkina Faso
 * OpenAPI spec version: 0.1.0
 */
import type { RestaurantStatusResultStatus } from './restaurantStatusResultStatus';

export interface RestaurantStatusResult {
  id: number;
  status: RestaurantStatusResultStatus;
}